Back-end Functionalities and Basic UI
after the patient logs in  appointment button will be displayed and the user can make an appointment, the appointment will be saved to the database and changed the status to "WAITING_APPROVAL". If the user type is superuser, superuser.ejs will be displayed and neccessary clinic info will be entered, this info will be saved to the db. I removed the loginoptions part, however, somehow I cannot direct the user to index page after logging in. also I translated texts to Turkish.